Rename multi-acc-helper to account manager.
