+++ tools/walletextension/accountmanager/account_manager_test.go
@@ -1,4 +1,4 @@
-package multiacchelper
+package accountmanager
 
 import (
 	"testing"
